♻️ Move tensor canonicalization out of QCO
Register QTensor scalarization through the native dialect hook and move
its regression tests to the QTensor suite. Remove the unused IR-to-utils
edge and restore ordinary one-way library dependencies.
