+++ mlir/lib/Dialect/QTensor/IR/QTensorOps.cpp
@@ -12,6 +12,8 @@
 
 #include "mlir/Dialect/QTensor/IR/QTensorDialect.h" // IWYU pragma: associated
 
+#include <mlir/Dialect/Arith/IR/Arith.h>
+
 // The following headers are needed for some template instantiations.
 // IWYU pragma: begin_keep
 #include <llvm/ADT/TypeSwitch.h>
