+++ mlir/lib/Dialect/QTensor/Utils/CMakeLists.txt
@@ -19,11 +19,9 @@ add_mlir_dialect_library(
   PRIVATE
   MLIRQTensorDialect
   MLIRFuncDialect
-  MLIRSCFDialect)
-
-# Keep the cyclic link dependency out of LLVM's object-library build ordering.
-# add_mlir_dialect_library already orders compilation after mlir-headers.
-target_link_libraries(MLIRQTensorUtils PUBLIC MLIRQCODialect)
+  MLIRSCFDialect
+  PUBLIC
+  MLIRQCODialect)
 
 mqt_mlir_target_use_project_options(MLIRQTensorUtils)
